Add and edit tests

+++ tests/clz/clz-64.c
@@ -2,7 +2,7 @@
 
 #define XLEN 64
 
-int HighestSetBit(int64_t x) {
+int64_t HighestSetBit(int64_t x) {
   for (int i = XLEN - 1; i >= 0; i--) {
     if ((x >> i) & 1) {
       return i;
